Move lookup table classes to utility and seal all algorithm classes
